Update in the test case for NSCBC.
The case direction is now specified in the input file so the code doesn't need to be recompiled.
The code is producing a segfault for the test in the y-direction when compiled with Intel and DEBUG=FALSE. The segfault is not reproduced with Intel and DEBUG=TRUE. Valgrind reported the problem coming from line 684 of NSCBC.H but that line is common to all directions so the same problem should occur in x and z. The validation in all directions was perfomed with DEBUG=TRUE